pyproject.toml
[tool.poetry]
name = "california-wild-fire-prediction"
version = "0.1.0"
description = ""
authors = ["shawn.kim <dun03191@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.11.5"
requests = "^2.32.3"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
poetry-plugin-dotenv = "^2.7.3"
ipykernel = "^6.29.5"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
folium = "^0.19.4"
rasterio = "^1.4.3"
geopandas = "^1.0.1"
numpy = "^2.2.4"
tqdm = "^4.67.1"
dbfread = "^2.0.7"
scikit-learn = "^1.6.1"
xgboost = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
